add function to add new script to the users database
